+++ src/plugins/console/public/application/containers/editor/editor.tsx
@@ -18,6 +18,7 @@ import {
 import { euiThemeVars } from '@kbn/ui-theme';
 
 import { i18n } from '@kbn/i18n';
+import { TextObject } from '../../../../common/text_object';
 
 import {
   EditorContentSpinner,
@@ -32,11 +33,13 @@ import {
   useServicesContext,
   useRequestReadContext,
   useRequestActionContext,
+  useEditorActionContext,
 } from '../../contexts';
 import type { SenseEditor } from '../../models';
 import { MonacoEditor, MonacoEditorOutput } from './monaco';
 import { getResponseWithMostSevereStatusCode } from '../../../lib/utils';
 
+const DEBOUNCE_DELAY = 500;
 const INITIAL_PANEL_WIDTH = 50;
 const PANEL_MIN_WIDTH = '100px';
 
@@ -58,6 +61,7 @@ export const Editor = memo(({ loading, setEditorInstance }: Props) => {
   } = useRequestReadContext();
 
   const dispatch = useRequestActionContext();
+  const editorDispatch = useEditorActionContext();
 
   const [fetchingMappings, setFetchingMappings] = useState(false);
 
@@ -69,6 +73,29 @@ export const Editor = memo(({ loading, setEditorInstance }: Props) => {
       subscription.unsubscribe();
     };
   }, []);
+
+  /* eslint-disable-next-line react-hooks/exhaustive-deps */
+  const debouncedUpdateLocalStorageValue = useCallback(
+    debounce((textObject: TextObject) => {
+      editorDispatch({ type: 'setCurrentTextObject', payload: textObject });
+    }, DEBOUNCE_DELAY),
+    []
+  );
+
+  // Always keep the currentTextObject in sync with the value in the editor
+  // to avoid losing the text object when the user navigates away from the shell
+  useEffect(() => {
+    // Only update the currentTextObject if the value has changed and is not empty
+    // This is to avoid setting the currentTextObject to the example text when
+    // the user clears the editor.
+    const textObject = {
+      ...currentTextObject,
+      text: inputEditorValue,
+      updatedAt: Date.now(),
+    } as TextObject;
+
+    debouncedUpdateLocalStorageValue(textObject);
+  }, [inputEditorValue, debouncedUpdateLocalStorageValue]);
 
   const [firstPanelWidth, secondPanelWidth] = storage.get(StorageKeys.WIDTH, [
     INITIAL_PANEL_WIDTH,
