+++ src/plugins/console/public/application/containers/editor/monaco/monaco_editor.tsx
@@ -6,7 +6,6 @@
  * Side Public License, v 1.
  */
 
-import { debounce } from 'lodash';
 import React, { CSSProperties, useCallback, useMemo, useRef, useState, useEffect } from 'react';
 import { EuiFlexGroup, EuiFlexItem, EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import { css } from '@emotion/react';
@@ -21,7 +20,6 @@ import {
   useRequestActionContext,
   useEditorActionContext,
 } from '../../../contexts';
-import { TextObject } from '../../../../../common/text_object';
 import {
   useSetInitialValue,
   useSetupAutocompletePolling,
@@ -39,8 +37,6 @@ export interface EditorProps {
   setValue: (value: string) => void;
 }
 
-const DEBOUNCE_DELAY = 500;
-
 export const MonacoEditor = ({ localStorageValue, value, setValue }: EditorProps) => {
   const context = useServicesContext();
   const {
@@ -51,7 +47,6 @@ export const MonacoEditor = ({ localStorageValue, value, setValue }: EditorProps
   const { toasts } = notifications;
   const {
     settings,
-    currentTextObject,
     restoreRequestFromHistory: requestToRestoreFromHistory,
   } = useEditorReadContext();
   const [editorInstance, setEditorInstace] = useState<
@@ -134,31 +129,6 @@ export const MonacoEditor = ({ localStorageValue, value, setValue }: EditorProps
   useSetupAutocompletePolling({ autocompleteInfo, settingsService });
 
   useSetupAutosave({ value });
-
-  /* eslint-disable-next-line react-hooks/exhaustive-deps */
-  const debouncedUpdateLocalStorageValue = useCallback(
-    debounce((textObject: TextObject) => {
-      editorDispatch({ type: 'setCurrentTextObject', payload: textObject });
-    }, DEBOUNCE_DELAY),
-    []
-  );
-
-  // Always keep the currentTextObject in sync with the value in the editor
-  // to avoid losing the text object when the user navigates away from the shell
-  useEffect(() => {
-    // Only update the currentTextObject if the value has changed and is not empty
-    // This is to avoid setting the currentTextObject to the example text when
-    // the user clears the editor.
-    if (currentTextObject?.text !== value && value !== '') {
-      const textObject = {
-        ...currentTextObject,
-        text: value,
-        updatedAt: Date.now(),
-      } as TextObject;
-
-      debouncedUpdateLocalStorageValue(textObject);
-    }
-  }, [value, currentTextObject, editorDispatch, debouncedUpdateLocalStorageValue]);
 
   // Restore the request from history if there is one
   const updateEditor = useCallback(async () => {
